+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -23,7 +23,7 @@ CHANGELOG
  * Added `cache:pool:prune` command to allow manual stale cache item pruning of supported PSR-6 and PSR-16 cache pool
    implementations
  * Deprecated `Symfony\Bundle\FrameworkBundle\Translation\TranslationLoader`, use 
-   `Symfony\Component\Translation\Loader\TranslationLoader` instead
+   `Symfony\Component\Translation\Reader\TranslationReader` instead
 
 3.3.0
 -----